==> notebooks/01_load_and_validate_raw_data.ipynb <==
from pathlib import Path
import requests 

##  Function to extract raw parquet data from the NYC taxi data 

def download_one_file_of_raw_data(year: int, month: int) -> Path:
    """"""
    URL = f"https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet"
    response = requests.get(URL)

    if response.status_code == 200:
        path = f"../data/raw/rides_{year}-{month:02d}.parquet"
        open(path, "wb").write(response.content)
        return path
    else:
        raise Exception(f"{URL} is not available")

download_one_file_of_raw_data(year=2022, month=1) 

import pandas as pd 

rides = pd.read_parquet("../data/raw/rides_2022-01.parquet")

rides.head(10)

### Make the rides dataframe only include the 2 columns pickup date time and pickup location id

rides = rides[["tpep_pickup_datetime", "PULocationID"]]

Rename the columns

rides.rename(columns={
    "tpep_pickup_datetime": "pickup_datetime",
    "PULocationID":"pickup_location_id"
    }, inplace = True)

rides.head(10)

### Here we see that there is a date from 2008 and a date from 2022 may, although this data is only supposed to be from jan 2022. let remove it...


rides["pickup_datetime"].describe()

### Making sure the data actually only consists of 2022 January data 

rides = rides[rides.pickup_datetime >= "2022-01-01"]
rides = rides[rides.pickup_datetime < "2022-02-01"]
rides["pickup_datetime"].describe()

### Store transformed file back to disk

rides.to_parquet("../data/transformed/validated_rides_2022_01.parquet")
