==> notebooks/02_transform_raw_data_into_ts_data.ipynb <==
import pandas as pd

### Reading in transformed data from 01

file_p = "../data/transformed/validated_rides_2022_01.parquet"
rides = pd.read_parquet(file_p)
rides.head(10)

Creating a new column pickup hour 


- .dt: This is an accessor object for datetime-like properties of the Series. It allows you to use various datetime-related attributes and methods on the pickup_datetime column.

- .floor("H"): The floor method is used to round down the datetime values to the nearest specified frequency. In this case, "H" specifies that you want to round down to the nearest hour. This means that for each datetime value in the pickup_datetime column, the minutes and seconds will be set to zero, effectively giving you the start of the hour for each datetime.

rides["pickup_hour"] = rides["pickup_datetime"].dt.floor("h")

Groupby to aggregate the pickup hour and location id

- .size(): This function calculates the size of each group. In this context, it counts the number of rides for each unique combination of pickup_hour and pickup_location_id. The result is a Series with a multi-index (composed of pickup_hour and pickup_location_id) and the count of rides as values.

- .reset_index(): This converts the multi-index Series back into a DataFrame. The pickup_hour and pickup_location_id become regular columns again, and the counts of rides become another column. By default, the column with the counts will be unnamed and indexed as 0.

rides.head(4)

### Explanation:

* For pickup_hour = 2022-01-01 and pickup_location_id = 4, there are 11 rides etc...

agg_rides = rides.groupby(["pickup_hour", "pickup_location_id"]).size()
agg_rides.head(5)

agg_rides = agg_rides.reset_index()
agg_rides

agg_rides.rename(columns={0:"rides"}, inplace=True)
agg_rides

There is no time where there is 0 pickups. However for other datasets this could be possible, therefore we will add a function which deals with this incase future data does have this to make sure we have a complete time-series and no missing values.

agg_rides["rides"].describe()

### Create a function which will make sure the number of rides matches up with pickup hour and pick up location id

The code defines a function add_missing_slots that ensures all hourly time slots have entries for each pickup location in the agg_rides DataFrame. It fills in any missing hourly time slots with zeros for the ride counts. 

agg_rides

agg_rides.info()

### Function to make sure that we remove pickup hour and location id for which there are empty rides. I.e. removing the gaps and replacing with 0 values for the time-series to ensure continuity. There are non-here fro 2022 data but may be the case with other parquet files. Also in other context this is vital e.g. e-commerce you may see this.

from tqdm import tqdm   

def add_missing_slots(agg_rides: pd.DataFrame) -> pd.DataFrame:
    
    location_ids = agg_rides['pickup_location_id'].unique()
    full_range = pd.date_range(
        agg_rides['pickup_hour'].min(), 
        agg_rides['pickup_hour'].max(), freq='h')
    output = pd.DataFrame()

    # Loop through the unique location id's and see it progress with tqdm
    for location_id in tqdm(location_ids):

        # For the current location_id, only rows corresponding to that location are extracted, 
        # retaining the pickup_hour and rides columns.
        agg_rides_i = agg_rides.loc[agg_rides.pickup_location_id == location_id, ['pickup_hour', 'rides']]
            
        # Quick way to add missing dates with 0 in a Series
        # Taken from https://stackoverflow.com/a/19324591
        agg_rides_i.set_index('pickup_hour', inplace=True)
        agg_rides_i.index = pd.DatetimeIndex(agg_rides_i.index)

        # The pickup_hour column is set as the index, and missing hours are added using reindex,
        # with the value 0 filling gaps in the rides column.
        agg_rides_i = agg_rides_i.reindex(full_range, fill_value=0)
        
        # After filling in missing hours, the pickup_location_id column 
        # is added back to preserve the association with the location.
        agg_rides_i['pickup_location_id'] = location_id

        output = pd.concat([output, agg_rides_i])
    
    # The index is reset, and the pickup_hour column is restored to its original name.
    output = output.reset_index().rename(columns={'index': 'pickup_hour'})
    
    return output

agg_rides_all_slots = add_missing_slots(agg_rides)


from typing import Optional, List
import plotly.express as px

def plot_rides(
    rides: pd.DataFrame,
    # locations: Optional[List[int]] means:
    # locations can either be a list of integers (e.g., [101, 102]) or None.
    # If omitted, it defaults to None, and the function will plot all locations.
    locations: Optional[List[int]] = None
    ):
    """
    Plot time-series data
    """

    # rides.pickup_location_id.isin(locations):
    # Checks whether the pickup_location_id in each row of the rides DataFrame is in the locations list.
    # Returns a boolean Series indicating True for rows that match.
    
    # rides[...]:
    # Selects rows where the condition is True

    # if locations else rides:
    # If locations is None (or an empty list), no filtering is applied, 
    # and the entire rides DataFrame is assigned to rides_to_plot 

    rides_to_plot = rides[rides.pickup_location_id.isin(locations)] if locations else rides

    fig = px.line(
        rides_to_plot,
        x="pickup_hour",
        y="rides",
        color='pickup_location_id',
        template='none',
    )

    fig.show()

### The plot is interactive if you click on it or highlight a focused area it will zoom in

plot_rides(agg_rides_all_slots, locations=[43])

# plot_rides(agg_rides_all_slots)

agg_rides_all_slots.to_parquet('../data/transformed/ts_data_2022_01.parquet')
