==> notebooks/08_lightgbm_model.ipynb <==
import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

# FRom 2022 29th Jan until 31st Dec
print(df['pickup_hour'].min())
print(df['pickup_hour'].max())

from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2022,6,1,0,0,0), # 2022-06-01 00:00:00 # 6 months of training data 
    target_column_name='target_rides_next_hour'
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

import lightgbm as lgb

# use only past rides data (there are other columns that you want to avoid thus just specifying them here)

past_rides_columns = [c for c in X_train.columns if c.startswith('rides_')]

X_train_only_numeric = X_train[past_rides_columns]

model = lgb.LGBMRegressor()
model.fit(X_train_only_numeric, y_train)

## Create Predictions

X_test_only_numeric = X_test[past_rides_columns]
predictions = model.predict(X_test_only_numeric)
predictions

## Evaluate 

from sklearn.metrics import mean_absolute_error

test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')
