==> notebooks/11_backfill_feature_store.ipynb <==
# Data from 22 to end of 23 now backfilled in feature store

from dotenv import load_dotenv
import os
from src.paths import PARENT_DIR

load_dotenv(PARENT_DIR / ".env")

HOPSWORKS_API_KEY = os.environ["HOPSWORKS_API_KEY"]
HOPSWORKS_PROJECT_NAME = "NYC_taxi_demand_project"

from datetime import datetime# Valid: Specify year, month, and day
dt = datetime.now().year 
#dt = dt.year # 2024
print(dt)

# Training data 

# All training data included from and to these dates
# 2022-01-29 00:00:00
# 2022-12-31 00:00:00

from datetime import datetime
import pandas as pd
from src.data import load_raw_data

from_year = 2022
to_year = dt # set to 2024 above takes all of 2024 available data
print(f'Downloading raw data from {from_year} to {to_year}')

rides = pd.DataFrame()
for year in range(from_year, to_year+1):
    
    # download data for the whole year
    rides_one_year = load_raw_data(year)
    
    # append rows
    rides = pd.concat([rides, rides_one_year])

print(f'{len(rides)=:,}')

## Data is from 2022 - 2023

# Might be worth simulating like 18 months because in feature pipeline 12 you simulate 52 weeks back

rides.tail(5)

print(rides['pickup_datetime'].dtype)

rides.head()

# # After kernel restart, load it back:
# import pandas as pd
# rides_2023_to_2024 = pd.read_pickle('rides_2023_to_2024.pkl')
# rides_2023_to_2024

#rides_2022_to_2023 = rides[(rides['pickup_datetime'] >= '2022-01-01') & (rides['pickup_datetime'] <= '2023-12-31')]
#rides_2022_to_2023

print(f'{len(rides)=:,}')

# 1. Check the earliest and latest dates
print("Earliest date:", rides['pickup_datetime'].min())
print("Latest date:", rides['pickup_datetime'].max())

# 2. Check how many rows you have
print("Number of rows:", len(rides))

# 3. Look at the first few and last few rows
print("\nFirst few rows:")
print(rides.head())
print("\nLast few rows:")
print(rides.tail())

# 4. Get a quick summary of the date range
print("\nDate range summary:")
print(rides['pickup_datetime'].describe())

from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)

# string to datetime
ts_data['pickup_hour'] = pd.to_datetime(ts_data['pickup_hour'], utc=True)

# add column with Unix epoch milliseconds
ts_data['pickup_ts'] = ts_data['pickup_hour'].astype(int) // 10**6

ts_data

print(ts_data['pickup_hour'].min())
ts_data['pickup_hour'].max()

import hopsworks

# hopsworks.logout()

# # Clear any existing connection
# project = None
# feature_store = None
# feature_group = None

project = hopsworks.login(
    project=HOPSWORKS_PROJECT_NAME,
    api_key_value=HOPSWORKS_API_KEY
)

feature_store = project.get_feature_store()

FEATURE_GROUP_NAME = 'time_series_hourly_feature_group'
FEATURE_GROUP_VERSION = 1

feature_group = feature_store.get_or_create_feature_group(
    name=FEATURE_GROUP_NAME,
    version=FEATURE_GROUP_VERSION,
    description="Time-series data at hourly frequency",
    primary_key = ['pickup_location_id', 'pickup_ts'],
    event_time='pickup_ts',
)#

print(f"Total rows in ts_data: {len(ts_data)}")

feature_group.insert(ts_data, write_options={"wait_for_job": False})
