==> notebooks/12_feature_pipeline.ipynb <==
import src.config as config

Earliest date: 2022-01-01 00:00:08


Latest date: 2023-12-31 23:59:59


Number of rows: 77964130

from datetime import datetime, timedelta,timezone
import pandas as pd

# Data in backfill feature store goes from 2022-01-01 00:00:00+00:00 to '2024-11-30 23:00:00+0000', tz='UTC')

# current_date = pd.to_datetime('2024-11-30 00:00:00+0000').floor('h')
from datetime import datetime, timezone
current_date = datetime.now(timezone.utc)

#current_date = pd.to_datetime(aware_utcnow()).floor('h')
print(f'{current_date=}')
print(type(current_date),f': {current_date=}')

# we fetch raw data for the last 28 days, to add redundancy to our data pipeline
fetch_data_to = current_date
fetch_data_from = current_date - timedelta(days=60)
fetch_data_from

# data from 2023-11-05 23:00:00 to '2023-12-03 23:00:00' 
# to fill the 28 days backfill 

# Problem arising with data dates here 

from src.data import load_raw_data

def fetch_batch_raw_data(from_date: datetime, to_date: datetime) -> pd.DataFrame:
    from_date_ = pd.Timestamp(from_date - timedelta(days=7*52)).replace(tzinfo=None)
    to_date_ = pd.Timestamp(to_date - timedelta(days=7*52)).replace(tzinfo=None)

    print(f'{from_date_=}, {to_date_=}')

    # Load data and normalize timezones for pickup_datetime
    rides = load_raw_data(year=from_date_.year, months=from_date_.month)
    rides['pickup_datetime'] = pd.to_datetime(rides['pickup_datetime'], errors='coerce').dt.tz_localize(None)
    rides = rides[rides.pickup_datetime >= from_date_]

    rides_2 = load_raw_data(year=to_date_.year, months=to_date_.month)
    rides_2['pickup_datetime'] = pd.to_datetime(rides_2['pickup_datetime'], errors='coerce').dt.tz_localize(None)
    rides_2 = rides_2[rides_2.pickup_datetime < to_date_]

    rides = pd.concat([rides, rides_2])

    # Shift the data to pretend this is recent data
    rides['pickup_datetime'] += timedelta(days=7*52)

    rides.sort_values(by=['pickup_location_id', 'pickup_datetime'], inplace=True)

    return rides

rides = fetch_batch_raw_data(from_date=fetch_data_from, to_date=fetch_data_to)
rides

print(rides['pickup_datetime'].min())
print(rides['pickup_datetime'].max())

from src.data import transform_raw_data_into_ts_data
ts_data = transform_raw_data_into_ts_data(rides)

print(ts_data['pickup_hour'].min())
print(ts_data['pickup_hour'].max())

# string to datetime
ts_data['pickup_hour'] = pd.to_datetime(ts_data['pickup_hour'], utc=True)

# add column with Unix epoch milliseconds
ts_data['pickup_ts'] = ts_data['pickup_hour'].astype(int) // 10**6

ts_data

print(ts_data['pickup_hour'].min())
print(ts_data['pickup_hour'].max())

import hopsworks

# connect to the project
project = hopsworks.login(
    project=config.HOPSWORKS_PROJECT_NAME,
    api_key_value=config.HOPSWORKS_API_KEY
)

# connect to the feature store
feature_store = project.get_feature_store()

# connect to the feature group
feature_group = feature_store.get_or_create_feature_group(
    name=config.FEATURE_GROUP_NAME,# "time_series_hourly_feature_group"
    version=config.FEATURE_GROUP_VERSION,
    description="Time-series data at hourly frequency",
    primary_key = ['pickup_location_id', 'pickup_ts'],
    event_time='pickup_ts',
)

print(f"Total rows in ts_data: {len(ts_data)}")

# This is only the last 28 days from the current date set at the top btw,  notebook 11 has taken all of the 2022-2024
feature_group.insert(ts_data, write_options={"wait_for_job": False})
