==> notebooks/14_inference_pipeline.ipynb <==
# Regulary fetch latest batch of features from feature store
# Current model in model registry

# inference pipeline - will compute new predicctions based on this data and 
# save it back to the feature store (predictions?)

%load_ext autoreload
%autoreload 2

from datetime import datetime, timedelta
import pandas as pd

current_date = pd.Timestamp.now(tz='UTC').floor('h')
current_date =  current_date - timedelta(days=180)

#current_date = pd.to_datetime('2024-10-30 23:00:00')
#current_date = pd.to_datetime('2024-10-02 23:00:00')
print(f'{current_date=}')

current_date

# TESTING RETURNING THE PICKUP_TS column
from src.inference import load_batch_of_features_from_store

features = load_batch_of_features_from_store(current_date)

features.columns

features

# # Make a copy of original dataframe and store pickup_ts separately
# original_features = features_with_pickup_ts_column.copy()
# pickup_ts_series = features_with_pickup_ts_column['pickup_ts'].copy()


# # Remove pickup_ts from features for predictions
# features_for_prediction = features_with_pickup_ts_column.drop('pickup_ts', axis = 1 )
# features_for_prediction

## Be aware that here we are making predictions without the 'pickup_ts' column 

from src.inference import (
    load_model_from_registry,
    get_model_predictions
)

model = load_model_from_registry()
predictions = get_model_predictions(model, features)

predictions

# pickup_hour = features_with_pickup_ts_column['pickup_hour'][0]
# pickup_hour

from src.inference import read_incremented_date

predictions['pickup_hour'] = current_date
predictions

## Add pickup_ts back into dataframe 

predictions['pickup_ts'] = predictions['pickup_hour'].astype(int) // 10**6
predictions

# 2 ways to try this: 

# (1) -  you have put in both pickup_hour and pickup_ts in here 
# and Pau's only has pickup hour. Try with this and then if not, try with pickup_ts
# (2) - 

from src.feature_store_api import get_feature_store
import src.config as config

# connect to the feature group
feature_group = get_feature_store().get_or_create_feature_group(
    name=config.FEATURE_GROUP_MODEL_PREDICTIONS, # model_predictions_feature_group
    version=4,
    description="Predictions generate by our production model",
    primary_key = ['pickup_location_id', 'pickup_ts'],
    event_time='pickup_ts',
)

feature_group.insert(predictions, write_options={"wait_for_job": False})
